+++ src/main/java/com/github/ttereshchenko/mailkit/settings/HeaderNamePrompter.java
@@ -1,0 +1,14 @@
+package com.github.ttereshchenko.mailkit.settings;
+
+import java.awt.Component;
+import java.util.List;
+import org.jetbrains.annotations.Nullable;
+
+@FunctionalInterface
+interface HeaderNamePrompter {
+
+    HeaderNamePrompter DEFAULT = new EditableComboHeaderNamePrompter();
+
+    @Nullable
+    String prompt(Component parent, List<String> suggestions);
+}